pyproject.toml
[tool.poetry]
name = "incremodoro"
version = "0.1.0"
description = ""
authors = ["Rickard Körkkö <rickard@rickardk.se>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
genanki = "^0.8.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
